+++ apipgen.config.ts
@@ -1,11 +1,10 @@
 import { defineConfig } from 'apipgen'
 
 const config = defineConfig({
-  pipeline: 'swag-ky-js',
+  pipeline: 'swag-axios-ts',
   input: {
     uri: 'https://petstore.swagger.io/v2/swagger.json',
   },
-  baseURL: 'aedwa',
   output: {
     main: 'dist-test/index.ts',
   },