+++ packages/swag-axios-js/src/parser/index.ts
@@ -5,6 +5,7 @@ import {
   parseHeaderCommits,
   parseMethodMetadata,
   parseMethodParameters,
+  transformBaseURL,
   transformDefinitions,
   transformParameters,
   transformUrlSyntax,
@@ -24,6 +25,10 @@ export function parser(configRead: ApiPipeline.ConfigRead) {
 
   const interfaces: StatementInterface[] = []
   const functions: StatementFunction[] = []
+
+  transformBaseURL(source, {
+    configRead,
+  })
 
   transformDefinitions(source.definitions, {
     interfaces,