+++ packages/swag-ky-ts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # apipgen-swag-ts-ky
 
+## 0.0.4
+
+### Patch Changes
+
+- chore: transform swag auto get baseURL
+- Updated dependencies
+  - @apipgen/swag-parser@0.1.2
+  - @apipgen/pipeline@0.1.2
+
 ## 0.0.3
 
 ### Patch Changes